# docs/openapi/climate-api-v1.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Climate Monitoring System API",
                 "description":  "Public API exposed by Climate.Gateway. Internal service endpoints are intentionally excluded.",
                 "version":  "v1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Local Docker Gateway"
                    }
                ],
    "paths":  {
...
                  "/api/events":  {
                                      "get":  {
                                                  "tags":  [
                                                               "Events"
                                                           ],
                                                  "parameters":  [
                                                                     {
                                                                         "name":  "riskType",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "$ref":  "#/components/schemas/RiskType"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "alertLevel",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "$ref":  "#/components/schemas/AlertLevel"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "sensorId",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "uuid"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "communityId",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "uuid"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "from",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "date-time"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "to",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "date-time"
                                                                                    }
                                                                     }
                                                                 ],
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "OK",
                                                                                "content":  {
                                                                                                "text/plain":  {
                                                                                                                   "schema":  {
                                                                                                                                  "type":  "array",
                                                                                                                                  "items":  {
                                                                                                                                                "$ref":  "#/components/schemas/EventResponse"
                                                                                                                                            }
                                                                                                                              }
                                                                                                               },
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "type":  "array",
                                                                                                                                        "items":  {
                                                                                                                                                      "$ref":  "#/components/schemas/EventResponse"
                                                                                                                                                  }
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "array",
                                                                                                                                 "items":  {
                                                                                                                                               "$ref":  "#/components/schemas/EventResponse"
                                                                                                                                           }
                                                                                                                             }
                                                                                                              }
                                                                                            }
                                                                            },
                                                                    "401":  {
                                                                                "description":  "Unauthorized"
                                                                            },
                                                                    "403":  {
                                                                                "description":  "Forbidden"
                                                                            }
                                                                },
                                                  "operationId":  "Events_get_api_events",
                                                  "security":  [
                                                                   {
                                                                       "Bearer":  [

                                                                                  ]
                                                                   }
                                                               ]
                                              }
                                  },
                  "/api/events/{id}":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "Events"
                                                                ],
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "id",
                                                                              "in":  "path",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "string",
                                                                                             "format":  "uuid"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "text/plain":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/EventResponse"
                                                                                                                                   }
                                                                                                                    },
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/EventResponse"
                                                                                                                                         }
                                                                                                                          },
                                                                                                     "text/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/EventResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                 }
                                                                                 },
                                                                         "401":  {
                                                                                     "description":  "Unauthorized"
                                                                                 },
                                                                         "403":  {
                                                                                     "description":  "Forbidden"
                                                                                 }
                                                                     },
                                                       "operationId":  "Events_get_api_events_id",
                                                       "security":  [
                                                                        {
                                                                            "Bearer":  [

                                                                                       ]
                                                                        }
                                                                    ]
                                                   }
                                       },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "AlertLevel":  {
                                                          "enum":  [
                                                                       "Green",
                                                                       "Yellow",
                                                                       "Orange",
                                                                       "Red"
                                                                   ],
                                                          "type":  "string"
                                                      },
...
                                       "RiskType":  {
                                                        "enum":  [
                                                                     "Flood",
                                                                     "Drought",
                                                                     "Storm",
                                                                     "Frost",
                                                                     "ForestFire"
                                                                 ],
                                                        "type":  "string"
                                                    },
                                       "EventResponse":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "id":  {
                                                                                           "type":  "string",
                                                                                           "format":  "uuid"
                                                                                       },
                                                                                "alertId":  {
                                                                                                "type":  "string",
                                                                                                "format":  "uuid"
                                                                                            },
                                                                                "sensorId":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "uuid"
                                                                                             },
                                                                                "communityId":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                },
                                                                                "riskType":  {
                                                                                                 "$ref":  "#/components/schemas/RiskType"
                                                                                             },
                                                                                "alertLevel":  {
                                                                                                   "$ref":  "#/components/schemas/AlertLevel"
                                                                                               },
                                                                                "description":  {
                                                                                                    "type":  "string",
                                                                                                    "nullable":  true
                                                                                                },
                                                                                "occurredAt":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "date-time"
                                                                                               },
                                                                                "resolvedAt":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "date-time",
                                                                                                   "nullable":  true
                                                                                               }
                                                                            },
                                                             "additionalProperties":  false
                                                         },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "description":  "Paste only the accessToken returned by POST /api/auth/login. Do not include the 'Bearer ' prefix; Swagger UI adds it automatically."
                                                          }
                                           }
                   },
...
}